Renamed file to iterative_binary_search and added comments and explanations
Refactored the binary search implementation:
- Renamed the file from binary_search.cpp to iterative_binary_search.cpp for clarity.
- Added inline comments and detailed explanations to improve code readability.
